fixed checkstyle errors
