+++ gateway/src/main/java/com/iluwatar/gateway/Gateway.java
@@ -4,5 +4,5 @@ package com.iluwatar.gateway;
  * Service interface.
  */
 interface Gateway {
-    void execute() throws Exception;
+  void execute() throws Exception;
 }
