+++ gateway/src/main/java/com/iluwatar/gateway/GatewayFactory.java
@@ -8,13 +8,13 @@ import java.util.Map;
  * example. It allows clients to register and retrieve specific gateways based on unique keys.
  */
 public class GatewayFactory {
-    private Map<String, Gateway> gateways = new HashMap<>();
+  private Map<String, Gateway> gateways = new HashMap<>();
 
-    public void registerGateway(String key, Gateway gateway) {
-        gateways.put(key, gateway);
-    }
+  public void registerGateway(String key, Gateway gateway) {
+    gateways.put(key, gateway);
+  }
 
-    public Gateway getGateway(String key) {
-        return gateways.get(key);
-    }
+  public Gateway getGateway(String key) {
+    return gateways.get(key);
+  }
 }
